% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_features}
\alias{filter_features}
\title{Filter Highly Correlated Features}
\usage{
filter_features(df, threshold)
}
\arguments{
\item{df}{A data frame with numeric features.}

\item{threshold}{Correlation threshold above which features are considered redundant.}
}
\value{
A data frame with reduced features.
}
\description{
Removes redundant features from a data frame by eliminating one feature in each highly correlated pair.
}
