% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_helpers.R
\name{generate_violin_plot}
\alias{generate_violin_plot}
\title{Generate a violin plot from clustered data}
\usage{
generate_violin_plot(data, feature, groupTreat = FALSE)
}
\arguments{
\item{data}{Data frame with cluster and feature columns}

\item{feature}{Feature to visualize}

\item{groupTreat}{Whether grouping by treatment}
}
\value{
A ggplot2 object
}
\description{
Generate a violin plot from clustered data
}
