% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ks_analysis.R
\name{perform_ks_analysis}
\alias{perform_ks_analysis}
\title{Perform KS test-based analysis between control and comparison matrix}
\usage{
perform_ks_analysis(ctrl_matrix, comp_matrix, features, cores = NULL)
}
\arguments{
\item{ctrl_matrix}{Matrix for control group}

\item{comp_matrix}{Matrix for comparison group}

\item{features}{Feature names}

\item{cores}{Number of cores to use for parallel computing}
}
\value{
A data.frame with KS statistics
}
\description{
Perform KS test-based analysis between control and comparison matrix
}
