% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_visualization.R
\name{plot_cluster_stability}
\alias{plot_cluster_stability}
\title{Plot cluster stability}
\usage{
plot_cluster_stability(stability, threshold = 0.6)
}
\arguments{
\item{stability}{Data frame from run_umap_leiden()$stability}

\item{threshold}{Numeric. Horizontal line marking stability threshold. Default 0.6.}
}
\description{
Plot cluster stability
}
