% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_visualization.R
\name{plot_clusters_proportions}
\alias{plot_clusters_proportions}
\title{Plot Clusters Proportions}
\usage{
plot_clusters_proportions(
  reduced_df,
  title = "Cell Proportions by Condition",
  save_path = NULL
)
}
\arguments{
\item{reduced_df}{A \code{data.frame} containing numeric features and
cluster assignments. By default, the function assumes that the last two
columns are \code{Clusters} and \code{Conditions}, and all preceding
columns are numeric features.}

\item{title}{Character. Plot title.}

\item{save_path}{Character or NULL. If provided, saves plot.}
}
\value{
ggplot object.
}
\description{
Creates stacked bar plot of cell proportions across conditions.
}
