% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_visualization.R
\name{plot_resolution_scan}
\alias{plot_resolution_scan}
\title{Plot resolution scan}
\usage{
plot_resolution_scan(resolution_scan)
}
\arguments{
\item{resolution_scan}{Data frame from run_umap_leiden()$resolution_scan}
}
\description{
Plot resolution scan
}
