% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_visualization.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP with flexible coloring (discrete or continuous)}
\usage{
plot_umap(
  umap_df,
  color_by = "Class",
  point_size = 1,
  discrete = NULL,
  low_color = "blue",
  mid_color = "white",
  high_color = "green",
  values_interval = NULL,
  midpoint_val = NULL
)
}
\arguments{
\item{umap_df}{Data frame from run_umap_leiden()$umap_df}

\item{color_by}{Character. Column name to color points by. Default "Class".}

\item{point_size}{Numeric. Size of points. Default 1.}

\item{discrete}{Logical or NULL.}

\item{low_color}{Color for minimum (continuous). Default "blue".}

\item{mid_color}{Color for midpoint (continuous). Default "white".}

\item{high_color}{Color for maximum (continuous). Default "green".}

\item{values_interval}{Numeric vector, lower and upper bounds of the color scale (default NULL).}

\item{midpoint_val}{Numeric, midpoint of the color scale (default NULL).}
}
\description{
Plot UMAP with flexible coloring (discrete or continuous)
}
