% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{rabanalyser.run}
\alias{rabanalyser.run}
\title{Run rabanalyser}
\usage{
rabanalyser.run()
}
\arguments{
\item{Null}{no paramters are needed}
}
\description{
function to run the rabanalyser shiny application.
}
\examples{
\dontrun{
rabanalyser.run()
}
}
\author{
Pernice Simone, Tealdi Simone
}
