% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_statistic.R
\name{two_sample_signed_ks_statistic}
\alias{two_sample_signed_ks_statistic}
\title{Compute Two-Sample Signed Kolmogorov-Smirnov Statistic}
\usage{
two_sample_signed_ks_statistic(sample1, sample2, ecdf1 = NULL)
}
\arguments{
\item{sample1}{A numeric vector.}

\item{sample2}{A numeric vector.}

\item{ecdf1}{An empirical cumulative distribution function (ECDF) for sample1.}
}
\value{
A numeric vector with the KS statistic and the signed KS statistic.
}
\description{
Calculates the signed and unsigned Kolmogorov-Smirnov statistic between two samples.
}
