% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleaning}
\alias{cleaning}
\title{Cleaning function to remove rows with missing values}
\usage{
cleaning(matrix)
}
\arguments{
\item{matrix}{A data frame or matrix.}
}
\value{
A cleaned data frame with complete rows.
}
\description{
Cleaning function to remove rows with missing values
}
